feat(testing): refactoring.
