--- testing/containers/database/config.go
@@ -1,9 +1,0 @@
-package database
-
-// Configuration represents database configuration settings
-type Configuration struct {
-	ConnectionString string
-	DatabaseName     string
-	DatabaseUser     string
-	Password         string
-}
