+++ testing/go.mod
@@ -1,35 +1,5 @@
 module go.zenithar.org/pkg/testing
 
-go 1.12
+go 1.14
 
-replace github.com/opencensus-integrations/gomongowrapper => github.com/Zenithar/gomongowrapper v0.0.2
-
-require (
-	github.com/Azure/go-ansiterm v0.0.0-20170929234023-d6e3b3328b78 // indirect
-	github.com/Microsoft/go-winio v0.4.12 // indirect
-	github.com/Nvveen/Gotty v0.0.0-20120604004816-cd527374f1e5 // indirect
-	github.com/cenkalti/backoff v2.1.1+incompatible // indirect
-	github.com/containerd/continuity v0.0.0-20181203112020-004b46473808 // indirect
-	github.com/dchest/uniuri v0.0.0-20160212164326-8902c56451e9
-	github.com/docker/go-connections v0.4.0 // indirect
-	github.com/docker/go-units v0.3.3 // indirect
-	github.com/gotestyourself/gotestyourself v2.2.0+incompatible // indirect
-	github.com/jackc/pgx v3.3.0+incompatible
-	github.com/jmoiron/sqlx v1.2.0
-	github.com/lib/pq v1.1.0
-	github.com/opencensus-integrations/gomongowrapper v0.0.1
-	github.com/opencontainers/go-digest v1.0.0-rc1 // indirect
-	github.com/opencontainers/image-spec v1.0.1 // indirect
-	github.com/opencontainers/runc v0.1.1 // indirect
-	github.com/ory/dockertest v3.3.4+incompatible // indirect
-	github.com/pkg/errors v0.8.1
-	github.com/sirupsen/logrus v1.4.1 // indirect
-	go.mongodb.org/mongo-driver v1.0.1
-	go.uber.org/zap v1.9.1
-	go.zenithar.org/pkg/db/adapter/postgresql v0.0.1
-	golang.org/x/sys v0.0.0-20190419153524-e8e3143a4f4a // indirect
-	golang.org/x/xerrors v0.0.0-20190410155217-1f06c39b4373
-	gopkg.in/ory-am/dockertest.v3 v3.3.4
-	gopkg.in/rethinkdb/rethinkdb-go.v5 v5.0.1
-	gotest.tools v2.2.0+incompatible // indirect
-)
+require github.com/ory/dockertest/v3 v3.6.0
